[MAKI-139] Структировала контроллер
