+++ maki-maki-ios/Manage Addresses/Adress.swift
@@ -12,41 +12,3 @@ struct Adress {
     var title: String
     var subTitle: String
 }
-
-struct Address: Codable {
-    var id: UUID
-    var street: String
-    var latitude: Double
-    var longitude: Double
-    var house: String
-    var flat: String
-    var type: Int
-}
-
-enum LocationType: String {
-    case home
-    case work
-    case other
-    
-    var image: UIImage {
-        switch self {
-        case .home:
-            return AppImage.homeMA.uiImage!
-        case .work:
-            return AppImage.workMA.uiImage!
-        case .other:
-            return AppImage.locationMA.uiImage!
-        }
-    }
-    
-    var description: String {
-        switch self {
-        case .home:
-            return "Home"
-        case .work:
-            return "Work"
-        case .other:
-            return "Other"
-        }
-    }
-}
