+++ maki-maki-ios/SelectLocation/Model/AddressService.swift
@@ -1,0 +1,24 @@
+//
+//  AddressService.swift
+//  maki-maki-ios
+//
+//  Created by Tami on 06.07.2023.
+//
+
+import Foundation
+
+class AddressService {
+    
+    func saveAddress() {
+    }
+    
+    func deleteAddress(){
+    }
+    
+    func fetchAddress() -> [Address]{
+        return []
+    }
+    
+    func updateAddress(){
+    }
+}
